backend/app/routes/protected.py
from fastapi import APIRouter, Depends, HTTPException
from ..utils.auth import oauth2_scheme
from ..utils.db import users_collection
from jose import jwt, JWTError
from ..config import settings

router = APIRouter()

async def get_current_user(token: str = Depends(oauth2_scheme)):
    try:
        payload = jwt.decode(token, settings.secret_key, algorithms=[settings.algorithm])
        email: str = payload.get("sub")
        if email is None:
            raise HTTPException(status_code=401, detail="Could not validate credentials")
    except JWTError:
        raise HTTPException(status_code=401, detail="Could not validate credentials")
    user = await users_collection.find_one({"email": email})
    if user is None:
        raise HTTPException(status_code=401, detail="User not found")
    return user

@router.get("/protected")
async def protected_route(current_user: dict = Depends(get_current_user)):
    return {"message": "This is a protected route", "name":current_user["first_name"], "email": current_user["email"], "company": current_user["company"]}